+++ packages/vscode-tailwindcss/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Detect classes in JS/TS functions and tagged template literals with the `tailwindCSS.classFunctions` setting ([#1258](https://github.com/tailwindlabs/tailwindcss-intellisense/pull/1258))
 - v4: Make sure completions show after variants using arbitrary and bare values ([#1263](https://github.com/tailwindlabs/tailwindcss-intellisense/pull/1263))
+- v4: Add support for `@source not` ([#1262](https://github.com/tailwindlabs/tailwindcss-intellisense/pull/1262))
+- v4: Add support for `@source inline(…)` ([#1262](https://github.com/tailwindlabs/tailwindcss-intellisense/pull/1262))
 
 # 0.14.9